requirements.txt
streamlit==1.35.0
google-generativeai==0.3.0
supabase==2.5.0
pdfplumber==0.11.0
pandas==2.2.0
openpyxl==3.1.2
plotly==5.22.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0
